workout_log/requirements.txt
Django
uuid
djangorestframework